PVS-50: Rework the type definitions to be by area
