+++ gql/types/index.js
@@ -1,0 +1,12 @@
+import userTypes from './user.js'
+import patchNoteTypes from './patchNotes.js'
+import faqTypes from './faqs.js'
+import { commonCategoryTypes, customCategoryTypes } from './categories.js'
+
+export default [
+  userTypes,
+  patchNoteTypes,
+  faqTypes,
+  commonCategoryTypes,
+  customCategoryTypes
+]
